examples/route_planner_agent/pyproject.toml
[project]
name = "route-planner-agent"
version = "0.1.0"
description = "Route planner agent with LlamaIndex, Couchbase vector store, and Capella AI model services"
authors = [
    {name = "Agent Catalog Team"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"

dependencies = [
    # LlamaIndex core and integrations
    "llama-index>=0.12.37,<0.13.0",
    "llama-index-vector-stores-couchbase>=0.4.0,<0.5.0",
    "llama-index-embeddings-openai>=0.3.1,<0.4.0",
    "llama-index-llms-openai-like>=0.3.5,<0.4.0",
    
    # Couchbase and database
    "couchbase>=4.0.0,<5.0.0",
    
    # Core dependencies
    "pydantic>=2.0.0,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    
    # Data processing and utilities
    "datasets>=2.0.0,<3.0.0",
    "requests>=2.28.0,<3.0.0",
    "httpx>=0.24.0,<1.0.0",
    
    # Agent Catalog dependencies
    "agentc-core @ ../../agent-catalog/libs/agentc_core",
    "agentc-llamaindex @ ../../agent-catalog/libs/agentc_integrations/llamaindex",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0,<8.0.0",
    "black>=23.0.0,<24.0.0",
    "ruff>=0.12.1,<0.13.0",
    "ipykernel>=6.29.5,<7.0.0",
    "jupyter>=1.0.0,<2.0.0",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "C90", "I", "N", "UP", "YTT", "S", "BLE", "FBT", "B", "A", "COM", "DTZ", "EM", "ISC", "ICN", "G", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SIM", "TID", "TCH", "ARG", "PTH", "PD", "PGH", "PL", "TRY", "NPY", "RUF"]
ignore = ["S101", "COM812", "ISC001"]

[tool.black]
line-length = 100
target-version = ['py312']